requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
langchain-google-genai>=0.0.1
langchain-community>=0.0.10
langchainhub>=0.1.0
python-dotenv>=1.0.0
litellm==1.30.0
requests>=2.31.0
pydantic>=2.0.0
psycopg2-binary>=2.9.9
supabase>=2.0.0
websockets>=13.0

# CLI & Utilities
typer>=0.9.0
rich>=13.0.0
openai>=1.0.0
anthropic>=0.7.0

# Server (optional but used in core)
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
